requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
torch==2.1.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
streamlit==1.29.0
plotly==5.18.0
stable-baselines3==2.2.1 